removed setHomeScreenBackground from webworks.tablet - #130
