+++ lib/ripple/platform/webworks.core/2.0.0/client/system.js
@@ -36,13 +36,6 @@ _self = {
 
     isMassStorageActive: function () {
         return transport.call(_uri + "isMassStorageActive");
-    },
-
-    setHomeScreenBackground: function (filePath) {
-        transport.call(_uri + "setHomeScreenBackground", {
-            get: {filePath: filePath},
-            async: true
-        });
     }
 };
