+++ lib/ripple/platform/webworks.handset/2.0.0/spec.js
@@ -216,7 +216,7 @@ module.exports = {
                     }
                 },
                 system: {
-                    path: "webworks.core/2.0.0/client/system",
+                    path: "webworks.handset/2.0.0/client/system",
                     feature: "blackberry.system",
                     children: {
                         event: {
